test/temp_consumer.py
Add a Clustering algorithm to Clustering stock using MLlib of Apache Spark in the code stream processing of Spark using Python: "import sys, json, hdfs, findspark, os
from pathlib import Path

path_to_utils = Path(__file__).parent.parent
sys.path.insert(0, str(path_to_utils))
sys.path.append("/app")

from confluent_kafka import Consumer, KafkaError
from script.utils import load_environment_variables
from pyspark.sql import SparkSession, Row
from pyspark.sql.window import Window
from pyspark.sql.functions import *
from pyspark.sql.types import *
from pyspark.sql.types import StructType, StructField, StringType, DoubleType, TimestampType
from datetime import datetime, timedelta

from InfluxDBWriter import InfluxDBWriter
from dotenv import load_dotenv
load_dotenv()
findspark.init()
env_vars = load_environment_variables()

KAFKA_TOPIC_NAME = "real-time-stock-prices"
KAFKA_BOOTSTRAP_SERVERS = "kafka1:29092,kafka2:29093,kafka3:29094"

# Configuration for Kafka Consumer
# conf = {
#     # Pointing to brokers. Ensure these match the host and ports of your Kafka brokers.
#     'bootstrap.servers': env_vars.get("KAFKA_BROKERS"),
#     'group.id': "myGroup",  # Consumer group ID. Change as per your requirement.
#     'auto.offset.reset': 'earliest'  # Start from the earliest messages if no offset is stored.
# }
# consumer = Consumer(conf)
# # Subscribe to the topic
# consumer.subscribe([env_vars.get("STOCK_PRICE_KAFKA_TOPIC"),])

scala_version = '2.12'
spark_version = '3.3.3'
packages = [
    f'org.apache.spark:spark-sql-kafka-0-10_{scala_version}:{spark_version}',
    'org.apache.kafka:kafka-clients:2.8.1'
]

if __name__ == "__main__":

    spark = (
        SparkSession.builder.appName("KafkaInfluxDBStreaming")
        .master("spark://spark-master:7077")
        .config("spark.jars.packages", ",".join(packages))
        .getOrCreate()
    )

    spark.sparkContext.setLogLevel("ERROR")

    stockDataframe = spark \
        .readStream \
        .format("kafka") \
        .option("kafka.bootstrap.servers", KAFKA_BOOTSTRAP_SERVERS) \
        .option("subscribe", KAFKA_TOPIC_NAME) \
        .load()

    stockDataframe = stockDataframe.select(col("value").cast("string").alias("data"))
    inputStream = stockDataframe.selectExpr("CAST(data as STRING)")

    stock_price_schema = StructType([
        StructField("stock", StringType(), True),
        StructField("date", TimestampType(), True),
        StructField("open", DoubleType(), True),
        StructField("high", DoubleType(), True),
        StructField("low", DoubleType(), True),
        StructField("close", DoubleType(), True),
        StructField("volume", DoubleType(), True)
    ])

    # Parse JSON data and select columns
    stockDataframe = inputStream.select(from_json(col("data"), stock_price_schema).alias("stock_price"))
    expandedDf = stockDataframe.select("stock_price.*")
    influxdb_writer = InfluxDBWriter('primary', 'stock-price-v1')
    #influxdb_writer = InfluxDBWriter(os.environ.get("INFLUXDB_BUCKET"), os.environ.get("INFLUXDB_MEASUREMENT"))
    print("InfluxDB_Init Done")

    def process_batch(batch_df, batch_id):
        realtimeStockPrices = batch_df.select("stock_price.*")
        for realtimeStockPrice in realtimeStockPrices.collect():
            timestamp = realtimeStockPrice["date"]
            tags = {"stock": realtimeStockPrice["stock"]}
            fields = {
                "open": realtimeStockPrice['open'],
                "high": realtimeStockPrice['high'],
                "low": realtimeStockPrice['low'],
                "close": realtimeStockPrice['close'],
                "volume": realtimeStockPrice['volume']
            }
            influxdb_writer.process(timestamp, tags, fields)

            # Convert Row to a dictionary
            row_dict = realtimeStockPrice.asDict()
            row_dict['date'] = row_dict['date'].isoformat()
            json_string = json.dumps(row_dict)
            print(json_string)
            print("----------------------")
            hdfs.write_to_hdfs(json_string)
        print(f"Batch processed {batch_id} done!")

    query = stockDataframe \
        .writeStream \
        .foreachBatch(process_batch) \
        .outputMode("append") \
        .start()

    query.awaitTermination()"
Let's think step by step.