requirements.txt
confluent-kafka
python-dotenv
yfinance
findspark
psycopg2-binary
influxdb_client
PyHDFS